Load text translation on init action for wp >=6.7.0
Fixes the notice message:

Translation loading for the plugin domain was triggered too early. This is usually an indicator for some code in the plugin or theme running too early. Translations should be loaded at the init action or later. Please see Debugging in WordPress for more information. (This message was added in version 6.7.0.)